Reorganize about:me to be less cluttered.
The about:me page is getting really long.

This changes the layout of the page so only one section of information
(activity, downloads, tabs, etc) will be displayed at once.

To add new content to the page with this change, create a new list item
in the menu for your content, formatted like the other list items.
Then wrap your new content in a div, and put it in the "boxborder" div,
making sure it's in the same order as the corresponding menu item.
aboutMe.js will take care of adding the onclick listener and the selection
what's displayed on the page.