Add geolocation to about:me
Adds a tab for providing geolocation services in about:me. For privacy reasons (and because using geolocation from an add-on doesn't have the notification bar asking for permission to access the geolocation service), about:me's default is to not run the geolocation until the user turns it on (there's a link in the tab to switch about:me's preference for it.

When the service is run, it calls getCurrentPosition(), which will return the latitude/longitude of the user.

It doesn't seem to work for Firefox 3.5, so unless I can find a way to access the service from there, you'll need to use 3.6 or 3.7. If the user has turned off the geo.enabled preference in Firefox, the extension will notify the user about it.

I can only assume it works with GPSD on Linux, as I don't have a compatible device to test it with.

XXX TODO XXX
Find some API and plot the coordinates on a map, displayed in the tab. (Can the Google Maps API be used on about: pages with a single API key?)